add chart and Currency Popover
