+++ src/pages/Beta/Swap/PairChart/styleds.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+import { Box } from '@pangolindex/components'
+
+export const ChartWrapper = styled(Box)`
+  width: 100%;
+  padding: 10px;
+  border-radius: 10px;
+  background-color: ${({ theme }) => theme.bg2};
+`
