+++ src/pages/Beta/Swap/WatchList/index.tsx
@@ -1,4 +1,4 @@
-import React, { useContext, useState } from 'react'
+import React, { useContext, useRef, useState } from 'react'
 import { Text, Box, Button } from '@pangolindex/components'
 import { ChainId } from '@pangolindex/sdk'
 import { Plus } from 'react-feather'
@@ -9,6 +9,10 @@ import WatchlistRow from './WatchlistRow'
 import { WatchListRoot, GridContainer } from './styleds'
 import Scrollbars from 'react-custom-scrollbars'
 import CoinChart from './CoinChart'
+import CurrencyPopover from './CurrencyPopover'
+// import { usePopper } from 'react-popper'
+import { useOnClickOutside } from 'src/hooks/useOnClickOutside'
+import useToggle from 'src/hooks/useToggle'
 
 const WatchList = () => {
   const { chainId = ChainId.AVALANCHE } = useActiveWeb3React()
@@ -16,16 +20,43 @@ const WatchList = () => {
   const theme = useContext(ThemeContext)
   const [selectedToken, setSelectedToken] = useState(coins[0])
 
+  const [open, toggle] = useToggle(false)
+  const node = useRef<HTMLDivElement>()
+  //  const [referenceElement, setReferenceElement] = useState<HTMLDivElement>()
+  // const [popperElement, setPopperElement] = useState<HTMLDivElement>()
+  const popoverRef = useRef<HTMLInputElement>(null)
+  const referenceElement = useRef<HTMLInputElement>(null)
+  // const { styles, attributes } = usePopper(referenceElement, popperElement, {
+  //   placement: 'auto',
+  //   strategy: 'fixed',
+  //   modifiers: [{ name: 'offset', options: { offset: [8, 8] } }]
+  // })
+
+  useOnClickOutside(node, open ? toggle : undefined)
+
   return (
     <WatchListRoot>
       <Box display="flex" alignItems="center" justifyContent="space-between">
         <Text color="text1" fontSize={32} fontWeight={500}>
           WatchList
         </Text>
-        <Box bgColor={theme.bg5 as any} p={'5px'}>
-          <Button variant="primary" backgroundColor="text8" color="text1" width={'32px'} height={'32px'} padding="0px">
-            <Plus size={12} color={theme.text1} />
-          </Button>
+        <Box bgColor={theme.bg5 as any} p={'5px'} ref={node as any}>
+          <Box ref={referenceElement} onClick={toggle}>
+            <Button
+              variant="primary"
+              backgroundColor="text8"
+              color="text1"
+              width={'32px'}
+              height={'32px'}
+              padding="0px"
+            >
+              <Plus size={12} color={theme.text1} />
+            </Button>
+          </Box>
+
+          {open && (
+            <CurrencyPopover getRef={(ref: HTMLInputElement) => ((popoverRef as any).current = ref)} coins={coins} />
+          )}
         </Box>
       </Box>
       <GridContainer>
