+++ src/pages/Beta/Swap/index.tsx
@@ -1,9 +1,10 @@
 import React from 'react'
-import { PageWrapper, ChartWrapper, GridContainer, TopContainer, StatsWrapper } from './styleds'
+import { PageWrapper, GridContainer, TopContainer, StatsWrapper } from './styleds'
 import MyPortfolio from './MyPortfolio'
 import WatchList from './WatchList'
 import PairStat from './PairStat'
 import Swap from './Swap'
+import PairChart from './PairChart'
 
 const SwapUI = () => {
   return (
@@ -11,7 +12,7 @@ const SwapUI = () => {
       <TopContainer>
         <StatsWrapper>
           <PairStat />
-          <ChartWrapper>Chart</ChartWrapper>
+          <PairChart />
         </StatsWrapper>
         <Swap />
       </TopContainer>
