+++ src/pages/Beta/Swap/styleds.tsx
@@ -23,10 +23,3 @@ export const GridContainer = styled(Box)`
   grid-gap: 12px;
   padding: 10px 0px;
 `
-
-export const ChartWrapper = styled(Box)`
-  width: 100%;
-  padding: 10px;
-  border-radius: 10px;
-  background-color: ${({ theme }) => theme.bg2};
-`
