+++ src/theme/styled.d.ts
@@ -25,6 +25,7 @@ export interface Colors {
   bg5: Color
   bg6: Color
   bg7: Color
+  bg8: Color
 
   modalBG: Color
   modalBG2: Color
